Move more registry support code out of core.
No real code changes here; just reorganization and import adjustments.

This required curating the symbols exported in registry/__init__.py
a bit more carefully, something I think we'll need to do more of.
