+++ python/lsst/daf/butler/core/__init__.py
@@ -12,7 +12,6 @@ from .constraints import *
 from . import ddl
 from .datasets import *
 from .datastore import *
-from .dbAuth import *
 from .exceptions import *
 from .fileDescriptor import *
 from .fileTemplates import *
