+++ python/lsst/daf/butler/registry/dbAuth.py
