+++ tests/test_oracle.py
@@ -29,7 +29,7 @@ import unittest
 import sqlalchemy
 
 from lsst.daf.butler import DimensionUniverse, ddl
-from lsst.daf.butler.core.registryConfig import RegistryConfig
+from lsst.daf.butler.registry import RegistryConfig
 from lsst.daf.butler.registry.databases.oracle import OracleDatabase
 from lsst.daf.butler.registry import Registry
 from lsst.daf.butler.registry.tests import DatabaseTests, RegistryTests
